HPC_FastAPI_Server/requirements.txt
fastapi
uvicorn[standard]
torch==2.1.2         
librosa
openai-whisper
nemo_toolkit[asr]



scipy
soundfile